api/requirements.txt
aiofiles==0.5.0
fastapi==0.54.1
uvicorn==0.11.4
requests==2.23.0
python-dotenv==0.13.0
databases[postgresql]==0.3.1
alembic==1.4.2